requirements.txt
aniso8601==8.1.0
ansi2html==1.6.0
arrow==0.17.0
asttokens==2.0.4
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Cython==0.29.21
executing==0.5.4
# Editable install with no version control (FaceMaskDetection==1.0)
#-e /cs/ghassan/scratch/mengliuz/SVN/FaceMaskDetection
Flask==1.1.2
Flask-RESTful==0.3.8
idna==2.10
importlib-metadata==3.3.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
littleutils==0.2.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mkl-fft
mkl-random
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603479632437/work
nvidia-ml-py3==7.352.0
olefile==0.46
packaging==20.8
pandas
Pillow @ file:///tmp/build/80754af9/pillow_1603822253009/work
pluggy==0.13.1
psutil==5.8.0
py==1.10.0
pycocotools==2.0.2
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six @ file:///tmp/build/80754af9/six_1605205313296/work
sorcery==0.2.1
tabulate==0.8.7
termcolor==1.1.0
toml==0.10.2
torch==1.4.0
torchaudio
torchvision==0.5.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.4.0
